Faster building of completion list, and also honor timeout
When completing, we now only go through the scandir list once, and
minimize the amount of calls to is_file() and is_dir() (these are
cached, and therefore fast, most of the time -- but not always,
e.g. on network filesystems)

This also improves the behavior of the timeout hint, allowing us to be
more responsive.